backend/requirements.txt
fastapi
uvicorn
bcrypt
python-jose
pydantic[email]
flask_jwt_extended